+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@ Release Notes
 ### Upcoming Changes
 
 - Jumi Actors is from this release onwards its own project with its own release cycle
+- Maven groupId was changed from `fi.jumi` to `fi.jumi.actors`
 
 ### Jumi Actors 0.1.196 (2012-09-19)